+++ src/config/index.js
@@ -1,7 +1,10 @@
-const getSlackCharsLimit = () => 39000;
-const getTeamsBytesLimit = () => 27000;
+const getSlackLimits = () => ({
+  chars: 30_000,
+  blocks: 50,
+});
+const getTeamsBytesLimit = () => 27_000;
 
 module.exports = {
-  getSlackCharsLimit,
+  getSlackLimits,
   getTeamsBytesLimit,
 };