+++ src/interactors/getReviewers/__tests__/index.test.js
@@ -1,13 +1,13 @@
 const input = require('./mocks/pullRequests');
 const getReviewers = require('../index');
 
+const getAuthors = (reviewers) => reviewers.map((r) => r.author.login);
+
 describe('Interactors | getReviewers', () => {
   it('groups reviews by author and calculate its stats', () => {
     const result = getReviewers(input);
     expect(result.length).toEqual(2);
-
-    const authors = result.map((r) => r.author.login);
-    expect(authors).toContain('manuelmhtr', 'jartmez');
+    expect(getAuthors(result)).toContain('manuelmhtr', 'jartmez');
 
     result.forEach((reviewer) => {
       expect(reviewer).toHaveProperty('author');
@@ -20,5 +20,11 @@ describe('Interactors | getReviewers', () => {
       expect(reviewer).toHaveProperty('stats');
       expect(reviewer.stats).toHaveProperty('timeToReview');
     });
+  });
+
+  it('excludes reviewers when the option is passed', () => {
+    const result = getReviewers(input, { excludeStr: 'manuelmhtr' });
+    expect(result.length).toEqual(1);
+    expect(getAuthors(result)).not.toContain('manuelmhtr');
   });
 });